add BTWebAuthenticationSessionClient to reduce duplicated logic
